essais pour régler le problème de modpow...

+++ Partie_1/tests_primalite.h
@@ -11,7 +11,7 @@ int is_prime_naive(long p);
 long modpow_naive(long a, long m, long n);
 
 /*Fonction calculant (a^m)%n*/
-int modpow(long a, long m, long n);
+unsigned long modpow(long a, long m, long n);
 
 /*retourne si a est un témoin de p (test de Miller)*/
 int witness ( long a , long b , long d , long p );
